+++ libs/application/templates/directorate-of-equality/equality-report/src/utils/constants.ts
@@ -36,6 +36,14 @@ export enum Gender {
 // it would fight the type system, so we don't.)
 export const UNKNOWN_DISPLAY_VALUE = 'Óþekkt'
 
+const DOE_NAMESPACE = 'DirectorateOfEquality'
+
+// Builds the `actionId` string the updateApplicationExternalData mutation expects,
+// from the same ApiActions enum the data providers and the service dispatch on —
+// a renamed action is then caught by the type checker at every call site.
+export const draftActionId = (action: ApiActions) =>
+  `${DOE_NAMESPACE}.${action}`
+
 export enum ApiActions {
   getCompanyData = 'getCompanyData',
   getDoeCompany = 'getDoeCompany',
